Replace "require" and "prefer" when publishing resolved versions
When publishing Gradle metadata, if a strategy says that we need
to use the resolved versions, then we will replace whatever "prefer"
or "require" version was set and publish the resolved version.

It will not, however, replace the "strictly" constraint, as it is
still possible to have a range in "strictly" and a "prefer" or
"require".